Always reuse existing branch
Remove the --reset-hard parameter and change the logic to always perform
'git reset $commit --hard'. Add a big note explaining how to go to the
previous commit if desired.

While using mrhlpr for some time now, I've noticed that I always want to
use the most recent commit from the merge request. Oftentimes I did not
notice, that I was working on an old, already checked out commit, and
this almost caused me to merge an outdated commit to master in two
occasions. Previously, we had a warning for working on such an outdated
branch, but it did not stand out enough, so I would not always notice
it.

I've also been using git-review to checkout gerrit patches, and noticed
that it always reuses branches without asking. So let's do this too, and
avoid some more human errors that way.